ml-predictor/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
pytest==7.4.3
 
